# contracts/public.openapi.yaml
openapi: 3.0.3
info:
  title: VOLA Public API
  version: "1.0.0"
  description: >
    Public API for the VOLA mobile and web clients. See
    docs/architecture/api-conventions.md for the conventions this spec
    follows (versioning, error shape, auth, status codes).


    Two behaviours are applied by response middleware and so are not
    per-operation choices, even though OpenAPI can only express them
    per-operation. Every GET returns a strong `ETag` and honours
    `If-None-Match` with a `304` — both are documented on each GET below.
    Every such response also carries `Cache-Control: private, no-cache` —
    "revalidate before reuse", which is the contract the `ETag` describes.
    A route that opts out sends `Cache-Control: no-store` and no validator
    (`/healthz` does, so a liveness probe is never answered `304`).


    And any response over ~1 KB is gzipped when the client sends
    `Accept-Encoding: gzip`; the `ETag` is computed from the *uncompressed*
    body, so it does not change with the encoding a client asks for.
servers:
  - url: http://localhost:8080/v1
    description: Local development
security:
  - ClerkBearerAuth: []
components:
  securitySchemes:
    ClerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued session token, verified server-side via Clerk's JWKS.
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - invalid_input
                - unauthorized
                - forbidden
                - not_found
                - already_exists
                - internal
            message:
              type: string
...
    Technique:
      type: object
      description: >
        One entry in the BJJ technique library. Deliberately not an Exercise:
        a technique isn't measured (you never log "3 sets of armbar"), and it
        lives in a graph — it comes from a position and is answered by
        counters. Reference content, identical for every user.
      required:
        [id, name, aliases, category, position, position_detail, gi_no_gi,
         typical_belt, description, when_to_use, setup_from,
         common_next_moves, common_counters, video_reference, source_notes,
         ibjjf_ruleset_id, created_at, updated_at]
      properties:
        id: { type: string, example: armbar-from-closed-guard }
        name: { type: string }
        aliases:
          type: array
          items: { type: string }
        category:
          type: string
          description: >
            Submission, Sweep, Pass, Escape, Takedown, Control/Pin, Transition,
            Guard Retention, Other. The colloquial label — what a coach says out
            loud. See `function` for the queryable axis.
          example: Submission
        function:
          type: string
          enum: [advance, reverse, escape, control, finish]
          description: >
            What the technique DOES, independent of where it happens. The five
            verbs of the taxonomy: advance (takedowns, passes, back takes),
            reverse (sweeps), escape (pin escapes, submission defence, guard
            retention), control (pins, rides, grip and frame systems), finish
            (submissions).

            Use this rather than `category` to ask "every way to X from here".
            `category` is colloquial and fuses two axes — "Takedown" means
            advance-at-standing and "Pass" means advance-at-guard-top — so
            "every way to advance" spans three categories while this is one
            value.

            ABSENT on a handful of movement fundamentals (breakfalls, grappling
            stance) which are library content rather than techniques and have no
            verb. Treat missing as "not applicable", not as an error.
          example: finish
        position:
          type: string
          description: Where it happens.
          example: Guard - Bottom
        position_detail: { type: string }
        gi_no_gi:
          type: string
          enum: [Both, "Gi Only", "No-Gi Only"]
        typical_belt:
          type: string
          example: White
          description: >
            The belt this is commonly taught from — an observation, NOT a
            recommendation and NOT a rule. It sits beside IBJJF legality and
            the two answer different questions: you may drill a technique long
            before you are allowed to compete it. Present it subordinate to
            `ibjjf`, never as a gate.
        description:
          type: string
          description: The mechanics — how the technique is performed.
        when_to_use:
          type: string
          description: >
            The decision — when the mechanics apply, and what has to be true
            first. Distinct from `description` on purpose; merging them
            produces a paragraph that answers neither question well.
        to_position:
          type: string
          description: >
            Where the technique LEAVES you. Closes the half of the graph
            `position` and `function` cannot express — where it starts and
            what it does, but not where it ends.


            SPARSE ON PURPOSE, and absent far more often than present. It is
            authored rather than derived: name parsing reaches 42% (mostly
            submissions, whose destination is the end of the exchange), and
            inverting `setup_from` infers a real position change for only 22
            of 466 — that edge links control-to-attack WITHIN a position.
            Populated where the destination is both knowable and useful
            (passes, sweeps, back takes, takedowns).


            ABSENT means NOT RECORDED, never "goes nowhere". A technique that
            genuinely leaves you where you started carries its own `position`
            value here, so "stays put" is a recorded fact rather than an
            absence. Do not infer a self-loop from a missing key.
          example: Side Control - Top
        setup_from:
          type: array
          description: >
            Graph edge — positions or techniques this is set up from. About 80%
            of entries name a real technique in the library.
          items: { type: string }
        common_next_moves:
          type: array
          description: >
            Graph edge — what commonly follows. Only ~29% of entries name a
            technique in the library; the rest is prose like "establish grips
            or inside ties". A client may link the ones that resolve but MUST
            render the others as plain text — a dead link is worse than text.
          items: { type: string }
        common_counters:
          type: array
          description: >
            Graph edge — what answers this technique. Only ~6% resolve to a
            library entry; treat as prose.
          items: { type: string }
        video_reference:
          type: string
          description: >
            Empty for every technique in the current library. Render nothing
            when blank rather than an empty section implying a missing video.
        source_notes: { type: string }
        ibjjf_ruleset_id:
          type: string
          description: >
            References an IBJJFRuleset. Empty if the technique predates the
            rulesets; treat that as "legality unknown", not "legal".
        ibjjf:
          allOf: [{ $ref: "#/components/schemas/IBJJFRuleset" }]
          description: >
            Resolved on the detail endpoint only, so a technique detail is one
            request. Absent from list responses — fetch /techniques/rulesets
            once and resolve `ibjjf_ruleset_id` locally.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    IBJJFRuleset:
      type: object
      description: >
        One IBJJF competition ruling, shared by every technique it applies to.
        25 of these cover all 466 techniques — normalised because `age_scope`
        has one distinct value library-wide and the most common `notes` string
        repeats 359 times at ~200 characters.
      required:
        [id, age_scope, rule_class, gi_allowed_belts, gi_note,
         no_gi_allowed_belts, no_gi_note, is_restricted, notes, sources]
      properties:
        id: { type: string, example: brown-black-only-adult }
        age_scope: { type: string, example: "Adult (18+) — Gi and No-Gi separated" }
        rule_class: { type: string, example: "Brown/Black only — Adult" }
        gi_allowed_belts:
          type: array
          description: >
            EMPTY MEANS "this division does not apply" — a no-gi-only technique
            has no gi belts — NOT "allowed at no belt". The reason is in
            `gi_note`.
          items: { type: string }
        gi_note: { type: string, example: "N/A — no-gi-specific" }
        no_gi_allowed_belts:
          type: array
          items: { type: string }
        no_gi_note: { type: string }
        is_restricted:
          type: boolean
          description: >
            Whether this is a GENUINE restriction rather than the shape of
            IBJJF's divisions. Adult no-gi has no white belt division, so a
            no-gi list of "Blue, Purple, Brown, Black" is the BASELINE.
            Do NOT re-derive this by comparing belt lists — that marks ~130
            ordinary techniques (pummelling, sit-outs) as restricted when the
            true number is 20. Use this field; it is computed at import.
        notes: { type: string }
        sources:
          type: array
          items: { type: string, format: uri }
...
    TechniqueWrite:
      type: object
      description: >
        The admin-writable surface. `id` and `source` are absent on purpose —
        the id is derived from the name at creation and immutable after, and
        the source is the server's.
      required: [name, category, position, gi_no_gi]
      properties:
        name: { type: string, example: São Paulo Pass }
        aliases: { type: array, items: { type: string } }
        category: { type: string, example: Pass }
        position:
          type: string
          description: Must be one of GET /admin/techniques/positions.
          example: Half Guard - Top
        position_detail: { type: string }
        gi_no_gi: { type: string, enum: [Both, Gi Only, No-Gi Only] }
        typical_belt: { type: string }
        description: { type: string }
        setup_from: { type: array, items: { type: string } }
        common_counters: { type: array, items: { type: string } }
        when_to_use: { type: string }
        common_next_moves: { type: array, items: { type: string } }
        video_reference: { type: string }
        source_notes: { type: string }
        ibjjf_ruleset_id: { type: string }
        function:
          type: string
          enum: [advance, reverse, escape, control, finish, ""]
        to_position:
          type: string
          description: Must be one of GET /admin/techniques/positions, or empty.
  headers:
    ETag:
      description: >
        Strong validator for this exact response body (a truncated SHA-256,
        base64url). Send it back as `If-None-Match` on the next request to get
        a 304 instead of the payload. Applied by response middleware, so it is
        present on every 200 from a GET, not just the ones listed here.
      schema: { type: string }
      example: '"9tK3xQwZ2sN4fRvB1cLmYg"'
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: >
        An `ETag` from a previous response. If it still matches, the server
        answers `304 Not Modified` with no body. Accepts a comma-separated
        list and `*`.
      schema: { type: string }
  responses:
    NotModified:
      description: >
        The representation is unchanged since the `ETag` in `If-None-Match`.
        No body, and deliberately no `Content-Length` or `Content-Type`.
      headers:
        ETag:
          $ref: "#/components/headers/ETag"
    BadRequest:
      description: Malformed request body or a value that failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, malformed, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated, but the caller isn't an admin.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The resource doesn't exist for this caller.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /techniques/{techniqueID}:
    get:
      summary: Get one technique
      operationId: getTechnique
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
        - name: techniqueID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The technique.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Technique" }
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /admin/techniques:
    get:
      summary: List the techniques the console can edit
      operationId: listAuthoredTechniques
      description: >
        Every technique with `source: "admin"` — what the console owns and can
        edit — and deliberately not the whole catalog. `PATCH` refuses a seeded
        row, because the seed JSON owns those and an edit would be reverted by
        the next deploy, so listing all of them would offer hundreds of rows of
        which a handful are actionable.


        This exists because the public `GET /techniques` cannot answer the
        question: its `Summary` carries no `source`, and adding one would put
        the string "seed" on every entry of every client's catalog to serve one
        admin screen.


        Unbounded. This set grows by hand, one technique at a time, and a
        console that silently truncated its own content would be worse than a
        slow one.
      responses:
        "200":
          description: >
            The authored techniques, ordered by id. `[]` when nothing has been
            authored yet — an empty array, never null.
          content:
            application/json:
              schema:
                type: object
                required: [techniques]
                properties:
                  techniques:
                    type: array
                    items:
                      $ref: "#/components/schemas/Technique"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      summary: Create a technique from the admin console
      operationId: createTechnique
      description: >
        Authoring the catalog without a deploy. The row is written with
        `source: "admin"`, which the seeder cannot touch — so a release never
        reverts it. An export command carries admin-authored content back into
        the seed JSON, which is how it reaches another environment.


        The **id is derived from the name** and is not accepted from the client:
        it becomes a foreign key in athletes' training records, so it outlives
        every other field and can never be changed. "São Paulo Pass" becomes
        `sao-paulo-pass`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechniqueWrite"
      responses:
        "200":
          description: The stored technique, including its generated id.
          content:
            application/json:
              schema:
                type: object
                required: [technique]
                properties:
                  technique:
                    $ref: "#/components/schemas/Technique"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: >
            The derived id is taken. Deliberately not an upsert — the id may
            already be referenced by somebody's training record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/techniques/{techniqueID}:
    patch:
      summary: Edit an admin-authored technique
      operationId: updateTechnique
      description: >
        Only `source: "admin"` rows. A seeded technique belongs to the deploy —
        editing it here would be reverted by the next release's re-seed, so it
        is refused with a 409 explaining that `techniques.json` is the way to
        change it. The id comes from the path and is never moved.
      parameters:
        - name: techniqueID
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechniqueWrite"
      responses:
        "200":
          description: The updated technique.
          content:
            application/json:
              schema:
                type: object
                required: [technique]
                properties:
                  technique:
                    $ref: "#/components/schemas/Technique"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The technique is seeded, so a deploy owns it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...